Move createdanceapi test helper to test helper folder
